Moved two classes
